convert trace_optimizer.py type hints

+++ znrnd/optimizers/trace_optimizer.py
@@ -25,6 +25,7 @@ Summary
 -------
 Module for the trace optimizer.
 """
+from __future__ import annotations
 from dataclasses import dataclass
 from typing import Callable
 
@@ -87,7 +88,7 @@ class TraceOptimizer:
             trace = np.trace(ntk)
 
             # Create the new optimizer.
-            new_optimizer = self.optimizer(float(self.scale_factor / (trace + eps)))
+            new_optimizer = self.optimizer(self.scale_factor / (trace + eps))
 
             # Create the new state
             new_state = TrainState.create(
